# eskip: add BenchmarkJsonUnmarshal